requirements.txt
fastapi 
uvicorn 
python-multipart 
langchain-openai 
python-dotenv